Update llnl/units and build conan recipe in cmake configure stage
There is some new (and old!) duplication between scipp and scippneutron
here. Most conan dependnecies of scippneutron are just needed by scipp,
so the conanfile and the cmake setup are identical. The new llnl-units
recipe is just another example.

For the recipe, the intention was actually to ship it with scipp, find
scipp, and use that recipe here. However, there is a chicken-egg
problem, since finding the scipp package has LLNL-Units as a dependency,
i.e., it cannot be fulfilled *after* calling find_package.

One solution could be to put conan setup into scipp-config.cmake.in and
ship the conanfile with scipp. Provided that this work, the only
duplication left would be that of the conan cmake function calls.

The disadvantages here would be that:
- Libraries using scip would be forced to use conan (not actuall issue
  for us now)
- scippneutron would "rely" and some dependencies of scipp getting
  install that are actually not runtime requirements, in particular
  GTest and pybind11.